pyproject.toml
[project]
name = "school-django-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-constance>=4.1.3",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
]
